+++ src/components/TeacherViews/LessonDownloads.view.tsx
@@ -22,6 +22,8 @@ import {
   getLessonDownloadsBreadCrumb,
   getBreadcrumbsForLessonPathway,
   getCommonPathway,
+  getBreadcrumbsForSpecialistLessonPathway,
+  getBreadCrumbForSpecialistDownload,
 } from "@/components/TeacherComponents/helpers/lessonHelpers/lesson.helpers";
 import { LessonPathway } from "@/components/TeacherComponents/types/lesson.types";
 import ResourcePageLayout from "@/components/TeacherComponents/ResourcePageLayout";
@@ -31,6 +33,7 @@ import { NextLesson } from "@/node-lib/curriculum-api-2023/queries/lessonDownloa
 import { useResourceFormState } from "@/components/TeacherComponents/hooks/downloadAndShareHooks/useResourceFormState";
 import { useHubspotSubmit } from "@/components/TeacherComponents/hooks/downloadAndShareHooks/useHubspotSubmit";
 import { LEGACY_COHORT } from "@/config/cohort";
+import { SpecialistLessonDownloads } from "@/node-lib/curriculum-api-2023/queries/specialistLessonDownload/specialistLessonDownload.schema";
 
 type BaseLessonDownload = {
   expired: boolean | null;
@@ -40,6 +43,8 @@ type BaseLessonDownload = {
   lessonCohort?: string | null;
   downloads: LessonDownloadsData["downloads"];
   hasDownloadableResources?: boolean;
+  isSpecialist: false;
+  developmentStageTitle?: string | null;
 };
 
 type CanonicalLesson = BaseLessonDownload & {
@@ -51,6 +56,8 @@ type NonCanonicalLesson = BaseLessonDownload & {
   nextLessons: NextLesson[];
 } & LessonPathway;
 
+type SpecialistLesson = SpecialistLessonDownloads["lesson"];
+
 type LessonDownloadsProps =
   | {
       isCanonical: true;
@@ -59,6 +66,10 @@ type LessonDownloadsProps =
   | {
       isCanonical: false;
       lesson: NonCanonicalLesson;
+    }
+  | {
+      isCanonical: false;
+      lesson: SpecialistLesson;
     };
 
 export function LessonDownloads(props: LessonDownloadsProps) {
@@ -69,10 +80,28 @@ export function LessonDownloads(props: LessonDownloadsProps) {
     downloads,
     hasDownloadableResources,
     expired,
+    isSpecialist,
   } = lesson;
+
   const commonPathway = getCommonPathway(
     props.isCanonical ? props.lesson.pathways : [props.lesson],
   );
+
+  const specialistPathway =
+    isSpecialist && !props.isCanonical
+      ? {
+          lessonSlug,
+          lessonTitle,
+          unitSlug: props.lesson.unitSlug,
+          programmeSlug: props.lesson.programmeSlug,
+          unitTitle: props.lesson.unitTitle,
+          subjectTitle: props.lesson.subjectTitle,
+          subjectSlug: props.lesson.subjectSlug,
+          developmentStageTitle: props.lesson.developmentStageTitle,
+          disabled: false,
+        }
+      : null;
+
   const {
     programmeSlug,
     keyStageTitle,
@@ -198,21 +227,35 @@ export function LessonDownloads(props: LessonDownloadsProps) {
       <MaxWidth $pb={80} $maxWidth={[480, 840, 1280]}>
         <Box $mb={isDownloadSuccessful ? 0 : 32} $mt={24}>
           <Breadcrumbs
-            breadcrumbs={[
-              ...getBreadcrumbsForLessonPathway(commonPathway),
-              getLessonOverviewBreadCrumb({
-                lessonTitle,
-                lessonSlug,
-                programmeSlug,
-                unitSlug,
-              }),
-              getLessonDownloadsBreadCrumb({
-                lessonSlug,
-                programmeSlug,
-                unitSlug,
-                disabled: true,
-              }),
-            ]}
+            breadcrumbs={
+              !isSpecialist
+                ? [
+                    ...getBreadcrumbsForLessonPathway(commonPathway),
+                    getLessonOverviewBreadCrumb({
+                      lessonTitle,
+                      lessonSlug,
+                      programmeSlug,
+                      unitSlug,
+                    }),
+                    getLessonDownloadsBreadCrumb({
+                      lessonSlug,
+                      programmeSlug,
+                      unitSlug,
+                      disabled: true,
+                    }),
+                  ]
+                : [
+                    ...getBreadcrumbsForSpecialistLessonPathway(
+                      specialistPathway,
+                    ),
+                    ...getBreadCrumbForSpecialistDownload({
+                      lessonSlug,
+                      programmeSlug,
+                      unitSlug,
+                      disabled: true,
+                    }),
+                  ]
+            }
           />
           <Hr $color={"grey60"} $mt={24} />
         </Box>